Add the morph feature
With morphs we more easily modify specific parts of the DOM directly
in the reflex. We also got access to a broadcaster which means that
we can use that to surgically alter certain DOM elements to our liking.

In this commit we also drop the commitment to maintain Channels 2.4.
Channels 2.4 should still work, but the CI showed weird issues and this
added to a maintainence headache which wasn't worth it
(unless explicitly asked for and paid by someone).