Make sure that reflex only renders context once
get_context_data should only be evaluated once per
reflex instance. The context is cached in a dictionary
if no context exists we can fill the dictionary with context

In the consumer we monkeypatch the view to use the reflex
get_context_data, that way we can make sure that the context
only ever gets evaluated once per reflex session.