object_list now works for ListView (#93)
Some views may have a 'setup' in the get part of
the view. For instance ListView expects that object_list
is added to self during the get request.

So if we get an AttributeError during the reflex we
execute the actual get request there, and then go on
to generate the context view.