Don't use exception, instead of figure out what view it is. (#124)
Figuring out which view it is and adding the correct property to
that view is better than trying to re-render the view itself.

This should solve for django's generic views. It won't work
for Django vanilla views package, but we can deal with this
later if someone show interest for that.